Created Welcome Page
I have created the welcome page, still needing to get the correct logo and start mapping out functionality of the buttons.
